requirements.txt
Babel==2.14.0
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==8.3.0
pillow==10.2.0
Pygments==2.17.2
rich==13.7.0
tkcalendar==1.6.1
ttkwidgets==0.13.0